texk/dvipdfm-x: No reason to use 'long' for 32-bit integers
git-svn-id: svn://tug.org/texlive/trunk/Build/source@38194 c570f23f-e606-0410-a88d-b1316a301751

+++ texk/dvipdfm-x/agl.h
@@ -32,7 +32,7 @@ struct agl_name {
   char *name;
   char *suffix;
   int   n_components;
-  long  unicodes[AGL_MAX_UNICODES];
+  int32_t  unicodes[AGL_MAX_UNICODES];
   struct agl_name *alternate;
   int   is_predef;
 };
@@ -40,15 +40,15 @@ typedef struct agl_name agl_name;
 
 extern char *agl_chop_suffix  (const char *glyphname, char **suffix);
 
-extern long  agl_sput_UTF16BE (const char *name,
+extern int32_t agl_sput_UTF16BE (const char *name,
 			       unsigned char **dstpp,
 			       unsigned char *limptr, int *num_fails);
 
 extern int   agl_get_unicodes (const char *glyphstr,
-			       long *unicodes, int max_uncodes);
+			       int32_t *unicodes, int max_uncodes);
 
-extern int   agl_name_is_unicode      (const char *glyphname);
-extern long  agl_name_convert_unicode (const char *glyphname);
+extern int      agl_name_is_unicode      (const char *glyphname);
+extern int32_t  agl_name_convert_unicode (const char *glyphname);
 
 extern const char *agl_suffix_to_otltag (const char *suffix);
